feat(screen-viewer): add screen viewer
Use screen-viewer to display the content for live-screen, preview-screen
and live-preview-screens
This represent the state of what will be dispayed on the secondary
screen. It is also used on the secondary screen(live-screen) to display
the contents

ScreenViewer is a component in the genenral directory and thus will
not include any direct electron integration.
This will make it reuseable across the primary and secondary screen.