Finally set text buffer.
In order to have an editable text_view in a list_view the textbuffer is
add at the list level, and attach to the text_view through RelmListItem bind
function. This way, any changes to the textview will update the external
textbuffer.

This allow for adding default text without overriding the updated
content when the bind function re-runs