Add new song
The basic implementation for adding new songs is complete.

New songs cannot have empty titles, but empty verses are allowed.
When editing is complete, verse strings are grabbed from their buffer and
converted into a string vector. Song title and verses are then converted to a
SongListItem and sent to the song list.

Todo:
- show focus verse on the activity viewer in the edit modal.