fix broken screen due to font scaling
The initial implementation caused broken screen on linux when if
maximizes.
I have replaced the `sqrt` fn with `cbrt` and a few changes to fix the
broken font scale.

I have also implemented `calculate_dyn_font_size` to get the proper
font-size for resized screen based on user selection (when the user
selection part is implemented)